+++ laspy/vlrs/known.py
@@ -8,9 +8,9 @@ import ctypes
 import logging
 import struct
 from typing import List, Optional, Any, Tuple, Dict, TypeVar, Type
+from copy import copy
 
 import numpy as np
-import pyproj
 
 from .vlr import BaseVLR, VLR
 from ..extradims import (
@@ -415,9 +415,27 @@ class WaveformPacketVlr(BaseKnownVLR):
 class GeoKeyEntryStruct(ctypes.LittleEndianStructure):
     _pack_ = 1
     _fields_ = [
+        # Id of the key
+        #
+        # Ids are broken down in sub domains:
+        # [    0,  1023]       Reserved
+        # [ 1024,  2047]       GeoTIFF Configuration Keys
+        # [ 2048,  3071]       Geographic/Geocentric CS Parameter Keys
+        # [ 3072,  4095]       Projected CS Parameter Keys
+        # [ 4096,  5119]       Vertical CS Parameter Keys
+        # [ 5120, 32767]       Reserved
+        # [32768, 65535]       Private use
         ("id", ctypes.c_uint16),
+        # Where to find the data for the key:
+        # 0 => The _actual_ value is stored directly in the "value_offset" member
+        # Otherwise, the tiff tag location is the record_id of the VLR in which the value is stored.
+        # In the case of LAS files the 2 possible values are `34736`, `34737`.
         ("tiff_tag_location", ctypes.c_uint16),
+        # Number of values in the key.
+        # Implied to be `1` if `tiff_tag_location` is 0
         ("count", ctypes.c_uint16),
+        # Depending on `tiff_tag_location`, this contains either
+        # the value itself _or_ the offset in the record_data of the containing VLR
         ("value_offset", ctypes.c_uint16),
     ]
 
@@ -487,82 +505,17 @@ class GeoKeyDirectoryVlr(BaseKnownVLR):
         b += b"".join(map(bytes, self.geo_keys))
         return b
 
-    def record_from_crs(self, crs: pyproj.CRS):
-        all_keys = {
-            "GTModelTypeGeoKey": GeoKeyEntryStruct(),
-            "GTRasterTypeGeoKey": GeoKeyEntryStruct(),
-            "GTCitationGeoKey": GeoKeyEntryStruct(),
-            "GeodeticCRSGeoKey": GeoKeyEntryStruct(),
-            "GeodeticCitationGeoKey": GeoKeyEntryStruct(),
-            "AngularUnitsGeoKey": GeoKeyEntryStruct(),
-            "ProjectedCRSGeoKey": GeoKeyEntryStruct(),
-            "LinearUnitsGeoKey": GeoKeyEntryStruct(),
-        }
-        all_keys["GTModelTypeGeoKey"].id = 1024
-        all_keys["GTModelTypeGeoKey"].count = 1
-        all_keys["GTRasterTypeGeoKey"].id = 1025
-        all_keys["GTRasterTypeGeoKey"].count = 1
-        all_keys["GTRasterTypeGeoKey"].value_offset = 1
-        all_keys["GTCitationGeoKey"].id = 1026
-        all_keys["GTCitationGeoKey"].tiff_tag_location = 34737
-        all_keys["GeodeticCRSGeoKey"].id = 2048
-        all_keys["GeodeticCRSGeoKey"].count = 1
-        all_keys["GeodeticCitationGeoKey"].id = 2049
-        all_keys["GeodeticCitationGeoKey"].tiff_tag_location = 34737
-        all_keys["AngularUnitsGeoKey"].id = 2054
-        all_keys["AngularUnitsGeoKey"].count = 1
-        all_keys["AngularUnitsGeoKey"].value_offset = 9102  # always degrees
-        all_keys["ProjectedCRSGeoKey"].id = 3072
-        all_keys["ProjectedCRSGeoKey"].count = 1
-        all_keys["LinearUnitsGeoKey"].id = 3076
-        all_keys["LinearUnitsGeoKey"].count = 1
-        added_keys = [
-            "GTModelTypeGeoKey",
-            "GTRasterTypeGeoKey",
-            "GTCitationGeoKey",
-            "GeodeticCitationGeoKey",
-            "AngularUnitsGeoKey",
-        ]
-
-        epsg = crs.to_epsg()
-        coord_name = crs.coordinate_operation.name + "|"
-        geodetic_name = crs.geodetic_crs.name + "|"
-        ascii_params = coord_name + geodetic_name
-
-        all_keys["GTCitationGeoKey"].count = len(coord_name.encode("utf8"))
-        all_keys["GeodeticCitationGeoKey"].count = len(geodetic_name.encode("utf8"))
-        all_keys["GeodeticCitationGeoKey"].value_offset = len(coord_name.encode("utf8"))
-
-        if crs.is_projected:
-            all_keys["GTModelTypeGeoKey"].value_offset = 1
-            all_keys["ProjectedCRSGeoKey"].value_offset = epsg
-            added_keys.append("ProjectedCRSGeoKey")
-        elif crs.is_geographic:
-            all_keys["GTModelTypeGeoKey"].value_offset = 2
-            all_keys["GeodeticCRSGeoKey"].value_offset = epsg
-            added_keys.append("GeodeticCRSGeoKey")
-
-        # assume linear units are same on both axes
-        all_keys["LinearUnitsGeoKey"].value_offset = int(crs.axis_info[0].unit_code)
-        added_keys.append("LinearUnitsGeoKey")
-
-        self.geo_keys = [all_keys[k] for k in added_keys]
-        self.geo_keys_header.number_of_keys = len(self.geo_keys)
-
     def parse_crs(self):
+        import pyproj
+
+        # TODO import is done here to avoid cyclic import,
+        # this should probably be fixed
+        from .geotiff import ProjectedCSTypeGeoKey, GeographicTypeGeoKey
+
         for key in self.geo_keys:
-            # get ProjectedCRSGeoKey by id
-            if key.id == 3072:
+            if key.id == ProjectedCSTypeGeoKey.id or key.id == GeographicTypeGeoKey.id:
                 return pyproj.CRS.from_epsg(key.value_offset)
         return None
-
-    @classmethod
-    def from_crs(
-        cls: Type[GeoKeyDirectoryType], crs: pyproj.CRS
-    ) -> GeoKeyDirectoryType:
-        self = cls()
-        self.record_from_crs(crs)
-        return self
 
     def __repr__(self):
         return "<{}({} geo_keys)>".format(self.__class__.__name__, len(self.geo_keys))
@@ -577,6 +530,10 @@ class GeoKeyDirectoryVlr(BaseKnownVLR):
 
 
 class GeoDoubleParamsVlr(BaseKnownVLR):
+    """
+    Stores all of the `double` valued GeoKeys.
+    """
+
     def __init__(self):
         super().__init__(description="GeoTIFF GeoDoubleParamsTag")
         self.doubles = []
@@ -611,23 +568,18 @@ class GeoDoubleParamsVlr(BaseKnownVLR):
 
 
 class GeoAsciiParamsVlr(BaseKnownVLR):
+    """
+    Stores all of the `ASCII` valued GeoKeys.
+
+    From GeoTIFF's spec:
+    To avoid problems with naive tiff dump programs the separator between geokeys is not
+    the null-terminator `\0` but `|`.
+
+    """
+
     def __init__(self):
         super().__init__(description="GeoTIFF GeoAsciiParamsTag")
         self.strings = []
-
-    def string(self, global_offset, size):
-        # The string in the vlr were stored contiguously, and null-separated
-        # we parsed them into list of strings.
-        # In the geokeys vlr, the offset to a string, is an offset in the
-        # contiguous strings, to this function does the conversion to find the
-        # correct string in our storage
-        count = 0
-        for string in self.strings:
-            count += len(string) + 1  # Account for the null byte
-            if count >= global_offset:
-                local_offset = global_offset - (count - len(string) - 1)
-                return string[local_offset:size]
-        raise IndexError(f"Invalid index: {global_offset}")
 
     def parse_record_data(self, record_data):
         self.strings = [s.decode("ascii") for s in record_data.split(NULL_BYTE)]
@@ -635,18 +587,6 @@ class GeoAsciiParamsVlr(BaseKnownVLR):
 
     def record_data_bytes(self):
         return NULL_BYTE.join(s.encode("ascii") for s in self.strings)
-
-    def record_from_crs(self, crs):
-        coord_name = crs.coordinate_operation.name + "|"
-        geodetic_name = crs.geodetic_crs.name + "|"
-        ascii_params = coord_name + geodetic_name
-        self.strings = [ascii_params, ""]
-
-    @classmethod
-    def from_crs(cls: Type[GeoAsciiParamsType], crs: pyproj.CRS) -> GeoAsciiParamsType:
-        self = cls()
-        self.record_from_crs(crs)
-        return self
 
     def __repr__(self):
         return "<GeoAsciiParamsVlr({})>".format(self.strings)
@@ -709,6 +649,8 @@ class WktCoordinateSystemVlr(BaseKnownVLR):
         return self._encode_string()
 
     def parse_crs(self):
+        import pyproj
+
         return pyproj.CRS.from_wkt(self.string)
 
     @staticmethod